add user auth to front and back

+++ mb-frontend/src/pages/UserIndex.jsx
@@ -5,25 +5,35 @@ import { utilService } from '../services/util.service.js'
 import { userService } from '../services/user.service.js'
 import { UserList } from '../cmps/UserList.jsx'
 import { UserFilter } from '../cmps/UserFilter.jsx'
+import { bugService } from '../services/bug.service.js'
 
 
 export function UserIndex() {
   const [users, setUsers] = useState([])
   const [filterBy, setFilterBy] = useState(userService.getDefaultFilter())
-  const debounceSetFilterBy = useCallback(utilService.debounce(onSetFilterBy,500), [])
+  const debounceSetFilterBy = useCallback(utilService.debounce(onSetFilterBy, 500), [])
+  const [loggedinUser, setLoggedinUser] = useState(userService.getLoggedinUser())
+
 
   useEffect(() => {
     loadUsers()
   }, [filterBy])
 
   async function loadUsers() {
-    console.log('loading users')
-    const users = await userService.query(filterBy)
+    if (!isAllowed()) return
+    const users = await userService.getUsers(filterBy)
     setUsers(users)
   }
 
   async function onRemoveUser(userId) {
     try {
+      const bugs = await bugService.query({ userId })
+      console.log(bugs)
+      if (bugs.length != 0) {
+        console.log('cannot remove user that has open bugs')
+        showErrorMsg('Cannot remove user')
+        return
+      }
       await userService.remove(userId)
       console.log('Deleted Successfully!')
       setUsers(prevUsers => prevUsers.filter((user) => user._id !== userId))
@@ -44,7 +54,6 @@ export function UserIndex() {
     }
     try {
       const savedUser = await userService.save(user)
-      console.log('Added User', savedUser)
       setUsers(prevUsers => [...prevUsers, savedUser])
       showSuccessMsg('User added')
     } catch (err) {
@@ -59,15 +68,18 @@ export function UserIndex() {
     try {
 
       const savedUser = await userService.save(userToSave)
-      console.log('Updated User:', savedUser)
       setUsers(prevUsers => prevUsers.map((currUser) =>
-      currUser._id === savedUser._id ? savedUser : currUser
+        currUser._id === savedUser._id ? savedUser : currUser
       ))
       showSuccessMsg('User updated')
     } catch (err) {
       console.log('Error from onEditUser ->', err)
       showErrorMsg('Cannot update User')
     }
+  }
+
+  function isAllowed() {
+    return loggedinUser?.isAdmin
   }
 
   function onSetFilterBy(filterBy) {
@@ -78,7 +90,8 @@ export function UserIndex() {
   function onChangePageIdx(pageIdx) {
     setFilterBy(prevFilter => ({ ...prevFilter, pageIdx }))
   }
-  if (!users) return <div>Loading...</div>
+  if (!isAllowed()) return <div> Only Admin can view this page</div>
+
   const isPaging = filterBy.pageIdx !== undefined
 
   return (
@@ -96,7 +109,7 @@ export function UserIndex() {
           </>}
         </div>
 
-        <UserFilter filterBy={filterBy} onSetFilterBy={debounceSetFilterBy} /> 
+        <UserFilter filterBy={filterBy} onSetFilterBy={debounceSetFilterBy} />
         <button onClick={onAddUser}>Add User </button>
         <UserList users={users} onRemoveUser={onRemoveUser} onEditUser={onEditUser} />
       </main>
